requirements.txt
# Core FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# AI and OCR
openai==1.51.2
# httpx is an indirect dependency of openai; pin a compatible version to avoid
# "Client.__init__() got an unexpected keyword argument 'proxies'" errors.
httpx==0.27.2
pillow==10.4.0
pytesseract==0.3.10

# PDF Generation
reportlab==4.2.2

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-mock==3.14.0
pytest-asyncio==0.23.7

# Database & Auth
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1
fastapi-users[sqlalchemy]==13.0.0
passlib[bcrypt]==1.7.4
